SQUID-935 added docs and tests
